requirements.txt
numpy
pillow
google-genai